requirements.txt
asgiref==3.3.4
Django==3.2.4
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-filter==2.4.0
django-multiselectfield==0.1.12
djangorestframework==3.12.4
et-xmlfile==1.1.0
flake8==3.9.2
future==0.18.2
gunicorn==20.1.0
jdcal==1.4.1
mccabe==0.6.1
Pillow==8.2.0
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
recipes==0.1
serializers==0.2.4
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
transliterate==1.10.2
workbook==1.1
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
